- refactor code - create NavigationBar - rename home view to mani view - create home view
