+++ lib/core/utils/app_router.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:fruit_market/core/utils/routes.dart';
-import 'package:fruit_market/features/home/presentation/views/home_view.dart';
-import 'package:fruit_market/features/home/presentation/views/widgets/home_view_body.dart';
+import 'package:fruit_market/main_view.dart';
+import 'package:fruit_market/features/home/presentation/views/widgets/main_view_body.dart';
 import 'package:fruit_market/features/login/presentation/views/create_account_view.dart';
 import 'package:fruit_market/features/login/presentation/views/login_view.dart';
 import 'package:fruit_market/features/onboarding/presentation/views/onboarding_view.dart';
@@ -17,7 +17,7 @@ abstract class AppRouter {
         builder: (context, state) {
           if (prefrenceObject != null) {
             if (prefrenceObject!.getBool('isLogin') == true) {
-              return const HomeView();
+              return const MainView();
             }
           }
           return const OnboardingView();
@@ -48,7 +48,7 @@ abstract class AppRouter {
       ),
       GoRoute(
         path: Routes.homeView,
-        builder: (context, state) => const HomeViewBody(),
+        builder: (context, state) => const MainViewBody(),
       ),
     ],
   );
